+++ User/User_App/app_roll_changing.h
@@ -7,10 +7,10 @@ typedef struct{
 	uint8_t state;	//信号状态
 	uint8_t delay_sw;		//延迟长度计算触发
 	float speed;
-	uint16_t length;	//延迟长度
+	float length;	//延迟长度
 	uint8_t step;	//换卷信号动作
-	uint16_t set_delay_length1; //设定延迟放电长度
-	uint16_t set_delay_length2; //设定电极闭合延迟长度
+	float set_delay_length1; //设定延迟放电长度
+	float set_delay_length2; //设定电极闭合延迟长度
 	
 	
 }roll_changingStruct_t;
